cli: Removed wlauto.utils.cli
Having the base arguments in wlauto.core.command make more sense
and wlauto.utils.cli didnt contain anything else.
